[BPK-4276] Restructure remaining B pages
