+++ docs/src/pages/BottomNavPage/AndroidBottomNavPage.js
