+++ docs/src/pages/BreadcrumbPage/WebBreadcrumbPage.js
@@ -1,0 +1,49 @@
+/*
+ * Backpack - Skyscanner's Design System
+ *
+ * Copyright 2016-2019 Skyscanner Ltd
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import React from 'react';
+import breadcrumbReadme from 'bpk-component-breadcrumb/README.md';
+import BpkBreadcrumb, { BpkBreadcrumbItem } from 'bpk-component-breadcrumb';
+
+import { WebComponentPage } from '../../components/ComponentPage';
+
+const components = [
+  {
+    id: 'default',
+    title: 'Default',
+    examples: [
+      <BpkBreadcrumb label="Example breadcrumbs">
+        <BpkBreadcrumbItem href="">Home</BpkBreadcrumbItem>
+        <BpkBreadcrumbItem href="">Page 1</BpkBreadcrumbItem>
+        <BpkBreadcrumbItem href="" active>
+          Page 2
+        </BpkBreadcrumbItem>
+      </BpkBreadcrumb>,
+    ],
+  },
+];
+
+const WebBreadcrumbPage = () => (
+  <WebComponentPage
+    examples={components}
+    readme={breadcrumbReadme}
+    packageName="bpk-component-breadcrumb"
+  />
+);
+
+export default WebBreadcrumbPage;
